+++ inject-java/src/test/groovy/io/micronaut/visitors/GeneratorTrigger.java
@@ -1,0 +1,11 @@
+package io.micronaut.visitors;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface GeneratorTrigger {
+}
